fix(routes): Allow RTS without specific rt entry
In order to return a message to the original sender (rts) the
sender's hostname:port was required to be in the route table
for the instance.  This might not always be the case as an
application may need to send messages to another only in an RTS
capacity and not as a message type routed message. Threfore the
requirement that an application host:port be listed in the route
table for this type of relationship isn't needed.

With this change, an application will be able to call rmr_rts_msg()
for any received message regardless of whether the sender is listed
in the local route table.

Change-Id: I9ab04a874524369dc218bacade0e34f7ea13ea95
Signed-off-by: E. Scott Daniels <daniels@research.att.com>