fix(build): Capture lib dir info on all flavours
For some flavours of Linux (arch, alpine, SuSE) the use of
lib64 as the installation directory made it difficult to
find the nano/nng libraries during the RMr build and install
process.  This change pulls the lib dir info as the nano/nng
CMake files do and thus is able to find the libraries
without needing a specific command line -D option on these
systems.

Change-Id: I70527865a47774ab446f6924788f5b864a776156
Signed-off-by: E. Scott Daniels <daniels@research.att.com>