test(app): Support lib64 based Linix flavours
Some flavours of Linux default to installing libraries
in lib64 rather than using lib as the default hardware
type and libXX as alternate architecture formats. These
changes allow the application oriented test scripts to
find the RMr libraries in the build directory regardless
of the choice of the underlying system.

Change-Id: I74aa982d7934f9c7dc7a5e86bbd1fa678c8dec61
Signed-off-by: E. Scott Daniels <daniels@research.att.com>