fix(send): Add second key lookup if sub-id is set
When finding an endpoint based on mtype+sub_id a second lookup
using just mtype is now performed if the mtype/sub_id combination
did not map to an endpoint.  This allows for "auto correction"
if the user application does not reset (clear) the sub_id when a
mtype should not be sent with one.

Change-Id: I5056e3d0a477d5a4272031ad8b2046e3ffca7c8a
Signed-off-by: E. Scott Daniels <daniels@research.att.com>