enhance(API): Add source IP support to msg header
The message header now carries both the hostname and IP
address of the source process.  The return to sender function
will use the IP address instead of the host name. It is possible
to set RMR_SRC_NAMEONLY=1 in the environment, or for the user
application to set this flag, to force only the use of the
hostname in the header.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Id6305fab79fabbffffb343af1094e48fbf662b5d