fix(rtc): Prevent core dump if \r used in rt file
Some operating systems, and/or filesystems, insist on using \r
as record termination (newline) indication rather than the
standard \n.  RMr was dependent on only finding \n as record
termination when parsing the route table entries, however as
it may be out of the user control to generate a static file with
the proper termination this change allows \r to be in the static
file.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ic18d5828f5a4b6855492b03718d105b57406431e