enhance(rtc): Tolerate missing newline in rtable
If the final record in a static route table does not have
a final newline (\n) this will be tolerated and the table
will be parsed as if the newline was present.  On some
systems it is not possible for the user to control when a
final newline is added (or not).

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I5627bd2a0b59a628a20757428df7960149c8229b