enhance(ci): Add ability to generate RPM package
The CMake configuration was enhanced to add the ability to
generate an RPM package in addition to a Debian package
provided the rpmbuild command is available in the current
environment.   The packages generated are now added to
the /tmp/build_output.yml file as fully qualified paths.

By default, the geneated package is a run-time package which
includes only the rmr libraries (no header files, and no
external libraries used to build RMr).  If the cmake
option -DDEV_PKG=1 is given at configure time, a dev package
will be generated which includes the header files, and
will have -dev added following the version info in the
package name (e.g. rmr-1.0.34.x86_64-dev.deb).

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ifb606e1f8b1ad0f08b3c4bd8b4012c0382947e04