Update patch set 6
Patch Set 6:

Ok, jenkins based vetting fixed.  @Chris  CMake default to generate run-time only package.   My assumption with the docker file change was that the package build/push uses a different docker file, so the change here does not affect that.  -1 this if that is not true and we'll need to chat.

Patch-set: 6
Label: Code-Review=0