Update patch set 7
Patch Set 7:

One last go round on this...  I tweaked the CMake to generate the static lib ONLY when generating a dev package (It finally dawned on me that .a is only valid in the dev.)  Changes to support enhanced package distribution will come in a different push.   With a +1 from @Chris, I think this is ready to merge.

Patch-set: 7
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=0