enhance(ci): Add multiple package build
The CI docker file now drives a CI build script which
will execute the unit and application based tests, and
if successful will build run-time and development .deb
and .rpm packages.  Packages are left in /tmp by default
and are indexed by build_packages.yml which is also left
in /tmp.

Change-Id: Ieaa97e1cab1d9cbb721e19d667192e336becec08
Signed-off-by: E. Scott Daniels <daniels@research.att.com>