Update patch set 4
Patch Set 4:

I built from this branch, copied out the 4 build artifacts, and tried to upload all to package-cloud. Failed bcos the dev and the non-dev versions both seem to have the same name internally, so the first upload wins and the second is rejected.  I unpacked them and saw that the control file in both has the same "Package: rmr" entry.  I think that needs to be different.

Patch-set: 4