Update patch set 6
Patch Set 6:

Thanks.  Adding Matti/Gueyoung for merge. 

The shift to amd64 seems to make sense. The label added by the build is automatic and system generated, so I'm not going to add anything which changes it -- don't want to go down that path.  If the package host wants to tweak for consistency on their side, then I'm OK with that. Appreciate the heads up, and _all_ of the eyes and testing.

Patch-set: 6
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=0