Update patch set 3
Patch Set 3:

This looks big, but I've just extracted the setup code from each doc source into a common module as I added the setup needed to generate markdown and plain ascii.  The only "real" chang is to setup.im and to the cmake file in the doc directory.

Patch-set: 3