Add publish script to the CI environment
The docker file will now set an entrypoint such that when the
container is run the ci/publish.ksh script is executed. The purpose
of the script is to copy the packages built into the image to a
mounted volume (/data by default). This eliminated the need for
an external process to extract files from the container.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I45ce98143bb84d2a802e1823ff003396bda6c9fe