Remove Nanomsg library generation
Nanomsg has reached end of life, and as NNG has proven to
be a viable replacement, with extra features, support for
nanomessage has been dropped. This change is to the CMake
configuration to remove the build and inclusion of libnng
from packages.

RMR libraries will also now be installed in the system
preferred library directory by default, and not blindly
into the lib directory.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ie7bba044f8e2e06fd7920ae581bc35d7e0e42677