Update patch set 3
Patch Set 3:

This change isn't as scary as it looks.  I rearranged the constants in numerical order to be more easily verified with respect to duplicates, so there it appears that there was a lot of churn.

Patch-set: 3
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=0