The change to fix a bug2 Fix return to sender retry bug; add env rts id
change 1: fix rts retry bug
The source IP address was not being restored properly after
a return to sender failure and subsequent retries were being
"reflected" rather than being sent to the message source.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I065e4b6b17896ee0de87e62916f72d2fc3067940

Support env var to define return to sender name

The host name in all environments cannot always be trusted to
be used as the message source from a "return to sender"
point of view.  This change allows the RMR_SRC_ID environment
variable to supply the "hostname" that is placed in a message
as the source and thus used when an application needs to
respond to the sender.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ie55c3147d6fc423c71315a43d23c88abc7e3c8d2

Tweak unit tests for better coverage

A warning was added if the multi-threaded receive process
drops messages. This tweak ensures it it tested during
unit test.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I380d58177cbc73eb1a10bbc52286922db7ead0b7

Bump version in CMake

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I2d4b19a5cfd1e2c9ecd196b0af2532788414e565