Add stretch module to python wrapper
The stretch module in the python bindings allows for
extensions and convenience functions. Initially this
contains a "receive all" function that collects all
queued messages and returns an array of message summaries
to the caller.

This change also modifies the expected constants test
such that it includes a check for the MTCALL flag, and
does not fail if RMR produces more constants than
expected and/or needed.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I862edd045c30bc4c81e13664acea6b91c229fb58