Update patch set 1
Patch Set 1:

Jenkins is failing to verify, but I think this is a jenkins test environment issue at the moment. Waiting on confirmation from Chris about that.   No hurry on the review as merge will likely be delayed until after the freeze is lifted.   My python skills are novice, so any better ways to do things will be appreciated.

Patch-set: 1