Update patch set 3
Patch Set 3:

(2 comments)

comments addressed, and ditched some trailing whitespace to keep gerrit calm.  Also addressed the unit test error.  The version of RMR in Jenkins doesn't export the MT CALL flag, and that caused the failure.  I think it's going to fail again in the test I just added (which is nice as I know the test is good).  I'll have to tweak that until we get 1.8 into jenkins.

Patch-set: 3
Label: Code-Review=0