Update patch set 2
Patch Set 2: Code-Review-1

I'm ok with the name, but we need to bump the version number in the CMake file unless we don't care that this change will fail with the attempt to push to package cloud.  I'm OK, with that, but if we really want to keep things clean we need to update the CMake file; which I think I prefer. 

Can you please update CMakeLists.txt  in the root directory and change the patch level to 1 (line 39 or there abouts).  We'll call this a bug fix.

Patch-set: 2