Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I think only the desired job was triggered here :)
> 
> Yes indeed it does appear that way -- WONDERFUL!  Thanks a) for the changes b) for pointing it out as I'd be very confused waiting for the second job :)

Well..... it failed because I didn't bump the RMR version number too and it's nattering on like:

ERROR: Committing multiple release files in the same commit OR rename/amend of existing files is not supported.

Which is fine after the merge, but this shouldn't prevent me from merging.

Builder didn't give a -1 as expected on the failure either, so something isn't quite right. 

Leaving this til morning at this point.

Patch-set: 1