Update patch set 2
Patch Set 2:

There is a test in rmr_mocks which ensures this (it tripped me up on the first submit and had to make it right).   Otherwise, it's impossible to generate a message with a nil body because of the way the rmr_allocate message works.  Please review and merge if good.  Going off for a bit, but will check back on this later this evening.

Patch-set: 2
Label: Code-Review=0