Correct inability to extend payload for rts msg
This change allows the application to increase the payload size
of an existing message buffer. This is needed when and application
must use the return to sender (rts) function to send a payload which
is larger than the payload in the received message.

Changes include a new RMR function: rmr_realloc_payload() and
changes to the python wrapper which will resize the payload if
the python application attempts to populate a message buffer which
does not have sufficent size.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>

Add python wrapper support for realloc payload

The python wrapper includes a direct call to the new RMR
extend payload funciton, and will automatically extend a
payload in the populate and set length function if the
current message isn't large enough to hold the bytes
being placed into the message.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>

Add man page for the new realloc payload function

New man page added, and references in other related pages
made as well.  A note relating to the need for the new
realloc function was also made in the return to sender
function's page.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Icc5a8e0bbe037f78f50b2bce3b4d700e030ab1b9