Prevent message truncation by populate function
An attempt to write a payload to a received message in prep
for returning the response to sender could result in the
response message being truncated if it is longer than the
payload in the received message. The populate and set length
function was modified to ensure that the target message has
a payload length which is not less than the message being
added; if it does the message payload is reallocated through
RMR.

This change requires RMR version 1.10.2 or later.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I8fa26eaace814344481c2df715c7c2486a38f6df