Update patch set 2
Patch Set 2:

[Chris added from an awareness point of view -- I think there is automation/validation processing that he wanted to watch.]

Please note that in order for this change to pass unit test the version in rmr-python/rmr-version.yaml was pushed to 1.10.2 which supports the underlying realloc_payload() function that allows a payload to be increased in size without affecting the message metadata needed for the return to sender function.   It means that if this change is merged, it will force an upgrade to RMR 1.10.2, so I've bumpped the version up to 2.0.0 as this is, I think, a breaking change from that point of view.

Patch-set: 2