Update patch set 2
Patch Set 2: Code-Review-1

Scott
I'm a bit torn here. On one hand, I do agree this is a semver major change (Though that is controversial, I actually have an open semver ticket talking about exactly this! https://github.com/semver/semver/issues/526)

On the other, breaking the API opens the door for us to shove other changes in at the same time, then cut a "release" of 2.0.0. That is, there might be other changes we might want to do but have been hesitant to do out of fear of breakage, and this might be a good opportunity to get some of those in.

Let me sleep on this and maybe on Monday I can take a look at rmr and see if there are other cleanups we should shove in at the same time; i'd be a shame to have to go to 3.0.0 shortly after and break "all of the users" =) yet again.

Our git branching model doesn't support this nicely. Ideally, we would have a branch off master, push a bunch of features into it, WITHOUT bumping the versions, and then finally "cut a release" by bumping the version all at once and merging into master. That is the famous Driessen model: https://nvie.com/posts/a-successful-git-branching-model/

So, I'm advocating we kinda shoehorn that in here.

Patch-set: 2
Label: Code-Review=-1