Minor version bump to move forward from trial ver
All changes required to fix bugs during the fall 2019
trial of RIC software will be applied to the previous
version number with only patch level version number changes.
This change advances the minor version number so as to
avoid any artifact naming conflict as the names don't
allow for disambiguation based on branch and/or release.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I8112c215e7508c5500bbb0b255ee141b354ed1d4