Documentation changes needed to support RTD
Changes which are necessary to support a docs directory that
is scraped for "read the docs."  Changes include:

Tweaks to manual pages to allow them to be combined into a
large "user's guide" (removed the final quit command).

Addition of an rtd subidrectory under source and the needed
.xfm source to generate the various .rst files in the top
level docs (plural).

Moved the top level readme to the main source directory as
it was out of place at the top; created a new top level README.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I708a750e3c58df4d0446eb3a5706c10bf65bed1d