Update patch set 3
Patch Set 3:

A lot of files touched just to get the mandated docs in.  Mostly allowing the manual pages to be reused and combined into a single user's guide.  The only really new information is the developers doc, which probably should be added to in the fullness of time.

Patch-set: 3
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=0