Update patch set 7
Patch Set 7:

The merge here failed, but the ultimate source of my confusion is that, it looked wrong both before and after the recent change to the jjb

it We want verify on all commits to any branch, but we want merge job on only merges to master. This wasn't the case before scott's commit, but it also wasn't the case prior, so this never really worked.

Patch-set: 7