Add SI95 transport support
The SI95 transport library provides an alternate transport
layer to NNG. The API presented by RMR on top of SI95
is exactly the same; only a link with rmr_si is needed to
make use of the alternate transport mechanism.

SI95 is NOT compatable with NNG and thus all RMR based
applications which must communicate must use the same
underlying transport library.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I0a971fa0f5de2dbb2fb7a95d5428e709b630f730