Update patch set 1
Patch Set 1:

Well yeah you were involved a bit Scott. On 11 November we had a big email discussion about tagging, verifying commits pushed to your b-temp branch, and so on.  In trying to configure the jobs for those needs I botched the merge job on the master branch.  

Thanks for the +1 on the repair.

Patch-set: 1