Update patch set 1
Patch Set 1:

in ONAP, a comment gets posted as a patch. That didn't happen until the job actually finished. So maybe it's not the case of the job not starting, but some commenting notification system isn't working in oran.

Submit a change to ONAP dcae and you will *immediately* see a comment posted as a patch saying verify runnng or whatever

Patch-set: 1