Correct documentation for MEID
The MEID acronym was incorrectly documented as "managed equipment"
and not "manage entity".  This change fixes this.

The RTD relase nodes were not being updated from the change log in
the repo root, and thus were not updated on the RTD site.  This
change adds the needed Makefile changes to ensure that all generated
docs for RTD are updated and published.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I2e3d741451221211a8abe49b7eca70dd6f935cbb