Add clarification to the rmr_send_msg man page
The need to ensure that the pointer returned by the rmr_send_msg()
function was not stressed in the manual page. This change adds
a caution to the man page to endure that the user programme
will not experience duplicate free or segment violations related
to reusing a stale pointer.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I0a26a482517314a29902313d9fe8b86b5b40a961