Fix large message bug in SI95 data callback
The receive buffer was not being allocated correctly and in some
cases would cause a segment fault when a large buffer was received.
The fix also includes dropping a message which is larger than the
allocated message.  The rmr_init() manual page has been updated to
better define the meaning of the max message length pararmeter which
is passed.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I20eb0d7ed8ba914b380807d9b3142d51d4f9f0b6