Fix session discconnect bug in interface to SI95
The disconnect of a session wasn't being properly recorded
and thus was preventing a reconnection attempt on the next
send attempt by the user application.

The warning about dropping messsages when the application cannot
keep up was also changed to be an error so that it is visible
at the default logging level.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ib07e4d0f9bf0d2b6436bb0b49be4ea37455509d3