Fix crash under SI95 with multiple receive threads
When the user application was using multiple threads to receive
messages it was alsmost certain to crash with a segfault.
This change corrects a bug in the ring handler which was
allowing a premature release of the ring lock.

A verification cookie has been added to the msg buffer to
facilitate validation of a message buffer when examining
crash output.

Issue-ID: RIC-218

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Iaa6519e415dcf7f8c52312e8e02d8cf4dca5866f