Add first set of SI95 unit tests and health check
This change adds the first series of unit tests for
the RMR SI95 code.  It also bundles the running of
the unit tests and the application level tests into
the CMake description such that "make test" will
run them all.

The health check application (health_ck) is now built
from the "support" source and added to the RMR run-time
package in /usr/local/bin.

Issue-ID: RIC-151

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I72804452afa5a972aa4ca0ded7b68131e16f5955