Revert RTS to use unidirectional connection
Because of potential performance impacts it was decided to
revert return to sender messages to NOT use the same
connection that the message was received on.  This applies
only to the SI95; RTS has always been on uni-directional
connections with NNG.

This change also enables the MEID routing in the SI95 code.

Issue-ID: RIC-153

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I6415effb3283f5961aa0098f1ff0d1380d7aa197