Fix message marker for src reset on send/rts
The SI95 receive function was not properly causing a
message buffer to be marked as "receive generated" and
if the msg buffer was subsequently used to send a message
(e.g. return to sender) the correct source address was
not placed into the message before sending.  This change
ensures that the message buffer received via SI95 are
marked so that a send will set the source address when
needed.

Issue-ID: RIC-227

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ic9c49834e5f066420463fa4568cfdf6f722f81cb