Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> i thought this was in response to verifying the new python code only
> 
> or do we have a separate problem where rmr itself is verifying when only python code is changed?? That sounds like a trigger problem in our CICD yamls.

Rmr runs 10 minutes worth of self test and unit test with each change, but those tests are driven only when RMR code changes.   One difference with those tests is that it doesn't depend on the virtual machine environment, so there has never been a need to bump the value in the .yaml file up.  The build environment is important to python as it needs the dependent version of RMR to be installed.  Has bumpping the version in the yaml file been tried (very hard for me to follow on gerrit disjoint messages).

Patch-set: 1