Update patch set 1
Patch Set 1:

sorry, I don't know how to reconcile your last comments.

As I understand, for python the rmr-version is relevent and anything going on in the main rmr repo is irrelevent. Is that the case or no? Stated alternatively, is the thing in package cloud bonked and used or no?

Patch-set: 1