Fix problem with RPM install
The additon of the health check supplemental application was
causing the RPM install to fail because of a misconception that
a library in the package wasn't already installed. This change
attempts to address this by hard linking the RMR library (.a)
into the supplemental applicaiton such that RPM won't assume
a depenendency.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: If9b70bed14a7994c517dc514c7963488301c8206