Correct the max receive message constant
The max receive size constant (a user programme only
constant) still reflected the original limit of 4k.
While internally this was not used and/or enforced, the
incorrect setting could cause some wrapper implementations
to incorrectly limit the max receive size.

This change sets this constant to the acknowledged max
receive size for all transports (64K).

Issue-ID: RIC-301

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I214b9d9dd67a00232e57bab8327dcb76f891b517