Rename health_ck support binary to rmr_probe
The health_ck name was determed to be too generic
and has been replaced with rmr_probe.  The application
performs the same health check function, and may at some
point in the futre be extended to provide other RMR
delivered probes.

This change also adds a useage message which was missing

Issue-ID: RIC-308

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I98f9dd7f9598b6a5682e4fb0d62c4146e8577e97