Eliminate the SI receive buffer length requirement
This change eliminates the need to enforce a receive
buffer maximum length as specified by the user application.
The length supplied will be assumed to be the "normal"
maximum and used as the default receive buffer size,
but when messages larger than this are received RMR will
now allocate a larger buffer to use.

Issue-ID: RIC-309

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I9dacadea40b982fde604b78789568b8118a89b35