Update patch set 4
Patch Set 4:

This is the change that you're holding the builder change on.  The version will be 3.8.0 as I realised that it's not just a patch to the last one out.  I don't know if you want to be poked on these changes or not but did because you're waiting on it.  I'll add Matti for merge, but I won't do that straight away as he'll review and merge out from under other reviewers.

Patch-set: 4
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>