Fix failure when in static only route table mode
When the RMR_RTG_SVC environment variable is set to -1, RMR
should enter static route table only mode and not attempt to
listen for dynamic updates from the route manager. There was
a misinterpretation of the environment variable name when
checking for this value which was causing the initialisation
to fail.

Man pages were also updated to remove references to the
deprecated RMR_MAX_RCV_BYTES constant. The constant was
marked as being deprecated.

Issue-ID: RIC-331

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I7b0b82397f3ad4f223682727213504d3fbdae518