Update patch set 2
Patch Set 2:

No hurry on this; I won't be looking at this til Monday.

This is the first of two doc changes.  This adds individual .rst files for the man pages. The content of the man pages was not changed, only the new .rst files generated into the docs directory so you shouldn't need to review those.   There is also a new page as requested for the rmr_get_const() function (in doc/src/man);  please do review that (there isn't much there).  

The changes to the code in the doc/src directory adds code to put in a doc title to the man pages as the .rst is generated (not a requirement for the roff version).  

Assuming this works, and the RTD site shows the table of contents for the man pages, and eash in its own file, I'll replace the user guide with the doc in the library as the next change set. (I don't want to yank what is there until the man pages show up as expected.)

Patch-set: 2
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>