Fix rmr_call() parameter checking bug
When RMR was extended to add wormhole support for
a rmr_wh_call() function, the main call code was broken
into an inward facing funciton.  A parameter check which
applies only to the outward facing rmr_call() API function
was accidentally moved to the internal funciton.  This change
moves the check back into the rmr_call() funciton.

Issue-ID: RIC-333

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Iab0c8cc33e8a9970c53b9248021a1273559f82f8