Add safe connect, fix possible seg fault in RTC
This change adds "safe connect" to the SI95 code which avoids
the connect to even port bug in Linux which could potentially
result in a successful connection when the remote process is
not running or listening on the attempted port.
Issue-ID: RIC-332

This change also includes a change to the level 2 debugging
in the route table collector (RTC) code which could have
resulted in a segment fault.
Issue-ID: RIC-335

Several of the man pages are updated with typo corrections.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ib6045f36850bd2a4b07a19bf11839ce3e9824b6a