Update patch set 4
Patch Set 4:

(8 comments)

Ok change set that fixes my fat fingering, addresses your additional comments, and has the correct user-guide.rst in the repo.   The diff on the user-guide is difficult; best just to review the raw new file as diff butchers it with what it matched in the original man page version.

Patch-set: 4