Update patch set 4
Patch Set 4:

No hurry on this.  I just wanted to wrap it up tonight, and likely I won't peek back for any comments until Monday.    This looks like a huge change, mostly because all of the .rst files rebuilt as a result of my tweaking the way definition lists are presented and to generate section headers with "annotation" that matches the length rather than the column width. There was also some white space and spelling cleanup that caused some of the source files to change.   The real meat of the change is the addition of the route tables guide (rt_tables.rst) which will deprecate one of the wiki pages when it hits RTD.   I pushed all of the .rst files through docutils to ensure that the resulting HTML looks good.

Patch-set: 4
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>