Update patch set 8
Patch Set 8:

I think I finally have all of the formatting glitches ironed out.  Disregard the comments about spacing (I'll update them directly).  It seems that whatever does the conversion to HTML for RTD doesn't respect the "no boarders" option for tables, and adds additional lines, so that was the ultimate cause.  I've worked round that and while it would be nice not to have any borders when requested, the definition lists look better and I think are easier to read.

Overview now points to the man.7 page; it was pulling that contents in, so for now it references it as the format wasn't pretty if someone was expecting a document. 

Document index should be consistent now, and header/titles are resolved; at least it seems that way when I pull the tox generated HTML.  

No hurry for a review, but it's ready.

Patch-set: 8