Update patch set 9
Patch Set 9:

(6 comments)

> Patch Set 8:
> 
> I don't much care for the stub overview.rst file.  Why not just nuke it and put its one sentence and one link directly about line 12 in index.rst:
> 
> 
> Please refer to the main RMR manual page for an overview of the library:
> 
> https://docs.o-ran-sc.org/projects/o-ran-sc-ric-plt-lib-rmr/en/latest/index.html

I will need to do this as a follow on change.  Tox  is barfing because the man page isn't there so the URL is not valid.   I guess I can't refer to something that will end up in the change.   I pulled from the index, and pulled the overview.rst.  All of the other minor tweaks are in at this point. 

I assume my second push (I forgot to pull the link from index with my first) will vet properly; if not I'll fix tomorrow.

Patch-set: 9