Add link to RTD index
This change adds the link to the main RMR manual page
in place of a formal overveiw document.

A few other small tweaks were made to the route table
document and glossary; the glossary code is shared so the
update triggers an update to the user guide .rst as well.

The overview source is deleted in this change.

Issue-ID: RIC-378

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I4972683faaab204dd0167b67a7510cb97e01d01a